requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
django-cors-headers>=4.4,<5.0
django-prometheus>=2.3,<3.0
drf-spectacular>=0.27,<1.0
celery[redis]>=5.4,<6.0
redis>=5.0,<6.0
psycopg[binary]>=3.2,<4.0
gunicorn>=22.0,<23.0
paramiko>=3.4,<4.0
pydantic>=2.9,<3.0
stripe>=10.0,<12.0
boto3>=1.35,<2.0
cryptography>=43.0,<44.0
python-decouple>=3.8,<4.0
